==> notebooks/vacancy_over_time/3.1 Join to foreclosure status.ipynb <==
# 3.1 Join to foreclosure status

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns
import math
import os
%matplotlib inline

def sound():
    os.system('afplay /System/Library/Sounds/Sosumi.aiff')

path = '/'.join(os.getcwd().split('/')[:-2])
print(path)

all_preds = pd.read_csv(path+'/outputs/all_predictions.csv', index_col=0, parse_dates=['date'])

# merge foreclosure status onto the predictions
fc_status = pd.read_csv(path+'/data/clean_data/parcel_status.csv')

fc_pivoted = fc_status.set_index(['parcel']).stack().reset_index().rename(columns={'level_1':'date',0:'fc_status'})
fc_pivoted['fc_status_mod'] = fc_pivoted['fc_status'] % 3
fc_pivoted = fc_pivoted.sort_values(by=['date','parcel'])
fc_pivoted = fc_pivoted.reset_index()
fc_pivoted = fc_pivoted.drop(fc_pivoted.columns[0],axis=1)

all_preds = all_preds.reset_index()
all_preds = all_preds.drop(all_preds.columns[0],axis=1)

all_preds[['fc_status','fc_status_mod']] = fc_pivoted[['fc_status','fc_status_mod']]

all_preds[['parcel','date','rf','lr','gbm','pv_count','fc_status','fc_status_mod']]\
    .to_csv(path+'/outputs/all_predictions_with_fc.csv')
